fix: tweak tooling for icons generation and rendering
